Fixed some errors with naming

+++ libft/libft.h
@@ -11,7 +11,7 @@ int		ft_putchar_fd(char c, int fd);
 int		ft_putstr_fd(char *s, int fd);
 int		ft_putnbr_fd(long n, int fd);
 size_t	ft_strlen(const char *c);
-char	*ft_itoa(int n);
+char	*ft_itoa(long n);
 int		ft_printf(const char *str, ...);
 
 typedef struct s_list
